+++ wa/src/lib.rs
@@ -27,7 +27,7 @@ fn get_handler() -> &'static FairMutex<native::Handler> {
         .expect("Backend has not initialized NATIVE_DISPLAY yet.") //|| Mutex::new(Default::default()))
 }
 
-fn set_display(id: u8, display: native::NativeDisplayData) {
+fn set_display(id: u16, display: native::NativeDisplayData) {
     let handler: &FairMutex<native::Handler> = get_handler();
     let mut d = handler.lock();
     d.insert(id, display);
@@ -46,7 +46,7 @@ pub mod window {
     use super::*;
     /// The current framebuffer size in pixels
     /// NOTE: [High DPI Rendering](../conf/index.html#high-dpi-rendering)
-    pub fn screen_size(id: u8) -> (f32, f32) {
+    pub fn screen_size(id: u16) -> (f32, f32) {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         (d.screen_width as f32, d.screen_height as f32)
@@ -54,7 +54,7 @@ pub mod window {
 
     /// The dpi scaling factor (window pixels to framebuffer pixels)
     /// NOTE: [High DPI Rendering](../conf/index.html#high-dpi-rendering)
-    pub fn dpi_scale(id: u8) -> f32 {
+    pub fn dpi_scale(id: u16) -> f32 {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         d.dpi_scale
@@ -62,7 +62,7 @@ pub mod window {
 
     /// True when high_dpi was requested and actually running in a high-dpi scenario
     /// NOTE: [High DPI Rendering](../conf/index.html#high-dpi-rendering)
-    pub fn high_dpi(id: u8) -> bool {
+    pub fn high_dpi(id: u16) -> bool {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         d.high_dpi
@@ -75,14 +75,14 @@ pub mod window {
     /// Window might not be actually closed right away (exit(0) might not
     /// happen in the order_quit implmentation) and execution might continue for some time after
     /// But the window is going to be inevitably closed at some point.
-    pub fn order_quit(id: u8) {
+    pub fn order_quit(id: u16) {
         let mut d = native_display().lock();
         let d = d.get_mut(id).unwrap();
         d.quit_ordered = true;
     }
 
     /// Shortcut for `order_quit`. Will add a legacy attribute at some point.
-    pub fn quit(id: u8) {
+    pub fn quit(id: u16) {
         order_quit(id)
     }
 
@@ -91,7 +91,7 @@ pub mod window {
     /// (for instance to show a 'Really Quit?' dialog box).
     /// If the event handler callback does nothing, the application will be quit as usual.
     /// To prevent this, call the function "cancel_quit()"" from inside the event handler.
-    pub fn request_quit(id: u8) {
+    pub fn request_quit(id: u16) {
         let mut d = native_display().lock();
         let d = d.get_mut(id).unwrap();
         d.quit_requested = true;
@@ -102,7 +102,7 @@ pub mod window {
     /// by calling "request_quit()". The only place where calling this
     /// function makes sense is from inside the event handler callback when
     /// the "quit_requested_event" event has been received
-    pub fn cancel_quit(id: u8) {
+    pub fn cancel_quit(id: u16) {
         let mut d = native_display().lock();
         let d = d.get_mut(id).unwrap();
         d.quit_requested = false;
@@ -113,21 +113,21 @@ pub mod window {
     /// NOTICE: on desktop cursor will not be automatically released after window lost focus
     ///         so set_cursor_grab(false) on window's focus lost is recommended.
     /// TODO: implement window focus events
-    pub fn set_cursor_grab(id: u8, grab: bool) {
+    pub fn set_cursor_grab(id: u16, grab: bool) {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         let _ = d.native_requests.send(native::Request::SetCursorGrab(grab));
     }
 
     /// Show or hide the mouse cursor
-    pub fn show_mouse(id: u8, shown: bool) {
+    pub fn show_mouse(id: u16, shown: bool) {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         let _ = d.native_requests.send(native::Request::ShowMouse(shown));
     }
 
     /// Show or hide the mouse cursor
-    pub fn set_window_title(id: u8, title: String) {
+    pub fn set_window_title(id: u16, title: String) {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         let _ = d
@@ -136,7 +136,7 @@ pub mod window {
     }
 
     /// Set the mouse cursor icon.
-    pub fn set_mouse_cursor(id: u8, cursor_icon: CursorIcon) {
+    pub fn set_mouse_cursor(id: u16, cursor_icon: CursorIcon) {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         let _ = d
@@ -145,7 +145,7 @@ pub mod window {
     }
 
     /// Set the application's window size.
-    pub fn set_window_size(id: u8, new_width: u32, new_height: u32) {
+    pub fn set_window_size(id: u16, new_width: u32, new_height: u32) {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         let _ = d.native_requests.send(native::Request::SetWindowSize {
@@ -154,7 +154,7 @@ pub mod window {
         });
     }
 
-    pub fn set_fullscreen(id: u8, fullscreen: bool) {
+    pub fn set_fullscreen(id: u16, fullscreen: bool) {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         let _ = d
@@ -163,35 +163,35 @@ pub mod window {
     }
 
     /// Get current OS clipboard value
-    pub fn clipboard_get(id: u8) -> Option<String> {
+    pub fn clipboard_get(id: u16) -> Option<String> {
         let mut d = native_display().lock();
         let d = d.get_mut(id).unwrap();
         d.clipboard.get()
     }
     /// Save value to OS clipboard
-    pub fn clipboard_set(id: u8, data: &str) {
+    pub fn clipboard_set(id: u16, data: &str) {
         let mut d = native_display().lock();
         let d = d.get_mut(id).unwrap();
         d.clipboard.set(data)
     }
-    pub fn dropped_file_count(id: u8) -> usize {
+    pub fn dropped_file_count(id: u16) -> usize {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         d.dropped_files.bytes.len()
     }
-    pub fn dropped_file_bytes(id: u8, index: usize) -> Option<Vec<u8>> {
+    pub fn dropped_file_bytes(id: u16, index: usize) -> Option<Vec<u8>> {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         d.dropped_files.bytes.get(index).cloned()
     }
-    pub fn dropped_file_path(id: u8, index: usize) -> Option<std::path::PathBuf> {
+    pub fn dropped_file_path(id: u16, index: usize) -> Option<std::path::PathBuf> {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         d.dropped_files.paths.get(index).cloned()
     }
 
     #[cfg(target_vendor = "apple")]
-    pub fn apple_view(id: u8) -> crate::native::apple::frameworks::ObjcId {
+    pub fn apple_view(id: u16) -> crate::native::apple::frameworks::ObjcId {
         let d = native_display().lock();
         let d = d.get(id).unwrap();
         d.view
